+++ models/listing.go
@@ -4,25 +4,25 @@ import "time"
 
 // Listing represents a Bnb listing
 type Listing struct {
-	Title           string
-	Price           float64
-	Currency        string // Currency symbol/code (฿, $, €, ₫, etc.)
-	Stars           float64
-	ReviewCount     int
-	URL             string
-	PageNumber      int    // Page number where this listing was found
-	AllPrices       []PriceInfo // For debugging: all prices found
-	
+	Title       string
+	Price       float64
+	Currency    string // Currency symbol/code (฿, $, €, ₫, etc.)
+	Stars       float64
+	ReviewCount int
+	URL         string
+	PageNumber  int         // Page number where this listing was found
+	AllPrices   []PriceInfo // For debugging: all prices found
+
 	// Detail page fields
-	IsSuperhost     bool
-	IsGuestFavorite bool
-	Bedrooms        int
-	Bathrooms       int
-	Beds            int
-	Description     string
-	HouseRules      string
+	IsSuperhost      bool
+	IsGuestFavorite  bool
+	Bedrooms         float64
+	Bathrooms        float64
+	Beds             float64
+	Description      string
+	HouseRules       string
 	NewestReviewDate *time.Time
-	Reviews         []Review
+	Reviews          []Review
 }
 
 // PriceInfo represents a price found in the listing
@@ -36,12 +36,8 @@ type PriceInfo struct {
 
 // Review represents a review for a listing
 type Review struct {
-	Date          time.Time
-	Score         float64
-	FullText      string
-	TimeOnAirbnb  string // How long the user has been on Airbnb
+	Date         time.Time
+	Score        float64
+	FullText     string
+	TimeOnAirbnb string // How long the user has been on Airbnb
 }
-
-
-
-
